Combine usage.py and usageplot.py into one tool.
